refactor: consolidate body capture into single file with IStartupFilter
Simplify the developer experience from 2 files + 2 lines to 1 file + 1 line:
- Merge options, middleware, and extension method into BodyCapture.cs
- Auto-register middleware via IStartupFilter (no app.UseMiddleware needed)
- builder.Services.AddHttpBodyCapture(builder.Configuration) is the only code change
- Fix otlphttp deprecation warning (renamed to otlp_http)
